Move get_series_name_from_files to helpers.collection
Refactor: extract get_series_name_from_files from app.py into helpers/collection.py and update usages. app.py now imports the helper; unit tests updated to import the function directly from helpers.collection (removed pytest.importorskip). No behavioral changes — preserves logic that strips parentheticals and trailing issue numbers (including "NNN of M") and falls back to DB name for missing/empty paths.
